Allow specifying a different network plugin.
For this we introduce a new mechanism for specifying string NCCL
parameters and change the name of the library where dlopen happens.

We also introduce a special behavior when passed an empty plugin name,
namely loading the plugin from the same executable that is running NCCL.
This is useful for the cases where the plugin and NCCL are statically
linked in the same binary. This is achieved by passing NCCL to dlopen,
which will cause it to return a handle to the main program.