allgatherv:
Improve performance of ncclGroup's of ncclBroadcast when each rank appears as a root in the group at most once.
NCCL_LAUNCH_MODE isn't listened to anymore. Now there's always an intra-process barrier, but only if there is at least one collective present in the ncclGroup. If the group is p2p-only, then we do comm-to-comm syncs (among the cpu threads) so that we still maximally prevent user CUDA API calls while we are launching kernels.
Improved perf of misaligned send/recv
Fix for group launches with multiple intraprocess cliques