Resolve some but not all deadlock when running with CUDA_LAUNCH_BLOCKING=1.
The two known sources of deadlock:

1) When the proxy thread is used (e.g. network peers): NCCL launches the
kernel before enqueueing work to the proxy thread since kernel launches
are higher latency. But, CUDA_LAUNCH_BLOCKING=1 causes that kernel launch
to block until the kernel completes, which the kernel can't do because
the proxy thread is blocked waiting for the work.

2) When managing multiple devices per OS thread: As NCCL has moved away
from the deprecated cudaLaunchCooperativeKernelMultiDevice, we now
loop through the devices owned by a thread and launch the kernels in
turn. Since the first launch will block on its kernel which is blocked
waiting for all the launches which communicate together, we deadlock.

This commit addresses only case 1. We simply make sure proxy work is
enqueued first when CUDA_LAUNCH_BLOCKING=1.