Fix char8_t typedef
Prior to this change the #if blocks were inside another #if branch
that limited them to #if __cplusplus, rendering the C part of the
feature test non-functional.  There was also a stray line that looked
like it came from the definition of nvtxExtPayloadTypeInfo.  In short,
it appears there has been a mis-merge.

Second, the standard versions used in the test were not correct.
- C: char8_t will be introduced in C23
  Source: https://gustedt.gitlabpages.inria.fr/c23-library/#char8_t-added-mandatory-change
  char16_t and char32_t were introduced in C11 (existing test was correct)
- C++: char8_t was introduced in C++20
  char16_t and char32_t were introduced in C++11
  Source: https://en.cppreference.com/w/User:D41D8CD98F/feature_testing_macros