[profiler example] Add missing "err.h" include in "profiler.h"
Several profiler_v*.h files uses ncclResult_t defined in err.h but the
header is not #included-ed in "profiler.h"

Currently the .c files in example does include it before including
"profiler.h" and therefore does not trigger compilation error.
But this commit should make the "profiler.h" header self-contained (as
it was in previous versions).